Fixes for OVMM execution stages (#228)
* fixes

* Fix ovmm execution

---------
