+++ src/stretch/agent/task/llm_plan/__init__.py
@@ -7,4 +7,4 @@
 # Some code may be adapted from other open-source works with their respective licenses. Original
 # license information maybe found below, if so.
 
-from .llm_plan_task import LLMPlanTask
+from .llm_plan_executor import LLMPlanExecutor
